+++ docs/scripts/header.js
@@ -1,5 +1,0 @@
-fetch("header.html").then(function (response) {
-  response.text().then(function (text) {
-    document.getElementById("header").innerHTML = text;
-  });
-});